Refactor the other tests to use the new helper
